+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 bootstrap5 extension Change Log
 2.0.2 under development
 -----------------------
 
+- Bug #6: yii\bootstrap5\BaseHtml::staticControl(): Argument #1 ($value) must be of type string, int given (dicrtarasov)
 - Bug #9: fixed default ActiveField::hintOptions (dicrtarasov)